requirements.txt
ecdsa==0.13
Fabric==1.11.1
gitdb==0.6.4
GitPython==2.0.5
nose==1.3.7
paramiko==1.17.0
pycrypto==2.6.1
smmap==0.9.0